refactor(chat): extract useDistinctModelsUsed and let OpenButton fold on a boolean
